// web/src/openapi/config/dasd.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * DASD config.
 */
export interface Config {
  /**
   * List of DASD devices.
   */
  devices?: Device[];
}
export interface Device {
  channel: string;
  /**
   * Specify target state of device. Either activate it or deactivate it.
   */
  state?: "active" | "offline";
  /**
   * If device should be formatted. If not specified then it format device only if not already formatted.
   */
  format?: boolean;
  /**
   * If device have set diag flag. If not specified then it keep what device has before.
   */
  diag?: boolean;
}